Use tolerance for linear function intersections
When calculating the intersection of two linear functions, the
`everywhere` and `nowhere` intersection types are now determined using a
small tolerance. In particular, if the slopes are within `1e-10` then
the linear functions either intersect `everywhere` or `nowhere` and if
the intercepts are within `1e-10` then the type is `everywhere`.