`PiecewiseConstant`s now have optionally restricted domains
`PiecewiseConstant` functions are now only defined over a pre-specified
subset of the real line. This is accomplished by including the domain
endpoints as the first and last entry of the boundary vector. This is
a cleaner solution than having a bunch of extra methods that can
restrict the domain of certain queries on piecewise constant functions.